generator/pyproject.toml
[tool.poetry]
name = "data-generator"
version = "0.1.0"
description = ""
authors = ["Fari Andana <fariqodri21@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.10"
kafka-python = "^2.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
